+++ test/protocol/grpc/grpc.rb
@@ -10,3 +10,4 @@ describe Protocol::GRPC do
 		expect(Protocol::GRPC::VERSION).to be =~ /\d+\.\d+\.\d+/
 	end
 end
+